Updated the Podcast Feed - a new trial account
Updated the Podcast Feed - a new trial account

Unfortunately the widget trial period has expired and not only the widget has stopped working AND the website has locked the widget so it won’t allow to copy the settings into a new (trial) account or view or edit the settings unless you purchase a plan. So I recreated the Podcast widget from scratch with another account.